requirements.txt
Flask
Gunicorn
textblob
google-generativeai